# Conflicts:
#	src/backend/BackgroundAgent/Installers/LogicInstaller.cs
#	src/backend/BackgroundAgent/Processing/Models/FileMetaInfo.cs

+++ src/backend/BackgroundAgent/Processing/Models/FileMetaInfo.cs
@@ -4,7 +4,7 @@
     {
         public string FileName { get; init; }
         
-        public long FileSize { get; init; }
+        public int FileSize { get; init; }
         
         public string FileType { get; init; }
